# Return VariableWidthBlock in slice direct reader