--- a/presto-spi/src/main/java/com/facebook/presto/spi/type/Chars.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/type/Chars.java
@@ -16,6 +16,7 @@ package com.facebook.presto.spi.type;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
 
+import static com.facebook.presto.spi.type.Varchars.byteCount;
 import static com.facebook.presto.spi.type.Varchars.truncateToLength;
 import static io.airlift.slice.SliceUtf8.countCodePoints;
 import static java.util.Objects.requireNonNull;
@@ -100,17 +101,30 @@ public final class Chars
     public static Slice trimTrailingSpaces(Slice slice)
     {
         requireNonNull(slice, "slice is null");
-        return slice.slice(0, sliceLengthWithoutTrailingSpaces(slice));
+        return slice.slice(0, byteCountWithoutTrailingSpace(slice, 0, slice.length()));
     }
 
-    private static int sliceLengthWithoutTrailingSpaces(Slice slice)
+    public static int byteCountWithoutTrailingSpace(Slice slice, int offset, int length)
     {
-        for (int i = slice.length(); i > 0; --i) {
+        requireNonNull(slice, "slice is null");
+        if (length < 0) {
+            throw new IllegalArgumentException("length must be greater than or equal to zero");
+        }
+        if (offset < 0 || offset + length > slice.length()) {
+            throw new IllegalArgumentException("invalid offset/length");
+        }
+        for (int i = length + offset; i > offset; i--) {
             if (slice.getByte(i - 1) != ' ') {
-                return i;
+                return i - offset;
             }
         }
         return 0;
+    }
+
+    public static int byteCountWithoutTrailingSpace(Slice slice, int offset, int length, int codePointCount)
+    {
+        int truncatedLength = byteCount(slice, offset, length, codePointCount);
+        return byteCountWithoutTrailingSpace(slice, offset, truncatedLength);
     }
 
     public static int compareChars(Slice left, Slice right)