SliceArrayBlock creates huge overhead for each Slice created. Slice
array sometimes can contain 1M entries which leads to about 56M
overhead. This causes full GC issues in production. Switch to
VariableWidthBlock to save memory.