@highker 2017-08-11
comments addressed