@highker 2017-08-17
@dain @nezihyigitbasi thanks for reviewing this patch; comments addressed