@highker 2017-08-17
Benchmark results on `HiveFileFormatBenchmark`; this is out of my expectation; we are getting even faster. @haozhun @dain 

*Without* the patch:
```
  read        VARCHAR_SMALL                   SNAPPY      HIVE_ORC                   9.88   149.7MB/s ±  2093.2kB/s ( 1.37%) (N = 150, α = 99.9%)
  read        VARCHAR_LARGE                   SNAPPY      HIVE_ORC                   9.88   146.6MB/s ±  2283.5kB/s ( 1.52%) (N = 150, α = 99.9%)
  read        VARCHAR_SMALL                   GZIP        HIVE_ORC                   10.76   134.9MB/s ±  2131.2kB/s ( 1.54%) (N = 150, α = 99.9%)
  read        VARCHAR_LARGE                   GZIP        HIVE_ORC                   10.76   137.6MB/s ±  4438.7kB/s ( 3.15%) (N = 150, α = 99.9%)
```

*With* the patch:
```
  read        VARCHAR_SMALL                   SNAPPY      HIVE_ORC                   9.88   152.7MB/s ±  1581.8kB/s ( 1.01%) (N = 150, α = 99.9%)
  read        VARCHAR_LARGE                   SNAPPY      HIVE_ORC                   9.88   156.0MB/s ±   956.2kB/s ( 0.60%) (N = 150, α = 99.9%)
  read        VARCHAR_SMALL                   GZIP        HIVE_ORC                   10.76   145.8MB/s ±   883.2kB/s ( 0.59%) (N = 150, α = 99.9%)
  read        VARCHAR_LARGE                   GZIP        HIVE_ORC                   10.76   146.0MB/s ±   793.2kB/s ( 0.53%) (N = 150, α = 99.9%)
```