@highker 2017-08-18
As discussed; we will keep the current implementation given there is no regression spotted.